+++ joyeriaSYS/frmCrearFactura.aspx.cs
@@ -485,12 +485,12 @@ namespace joyeriaSYS
             return nuevaFactura.idFactura;
         }
 
-        private string EstadoFacturaEnLetras(byte num)
+        private string EstadoFacturaEnLetras(int num)
         {
-            if(num == Convert.ToByte(EstadoFacturas.EnCreacion))
+            if(num == Convert.ToInt32(EstadoFacturas.EnCreacion))
             {
                 return "En creación";
-            }else if (num == Convert.ToByte(EstadoFacturas.Finalizada))
+            }else if (num == Convert.ToInt32(EstadoFacturas.Finalizada))
             {
                 return "Finalizada";
             }
